test: add executable InstanceStart e2e
